# Conflicts:
#	app/pom.xml
#	app/src/main/java/nl/scalda/pasimo/datalayer/testdao/TestDAOTeacher.java
#	app/src/main/java/nl/scalda/pasimo/model/employeemanagement/EducationTeam.java
#	app/src/main/java/nl/scalda/pasimo/service/EducationTeamService.java

+++ app/src/main/java/nl/scalda/pasimo/model/employeemanagement/EducationTeam.java
@@ -1,5 +1,6 @@
 package nl.scalda.pasimo.model.employeemanagement;
 
+import java.io.Serializable;
 import nl.scalda.pasimo.datalayer.factory.DAOFactory;
 import nl.scalda.pasimo.datalayer.mysqldao.MYSQLDAOCoachGroup;
 import nl.scalda.pasimo.datalayer.testdao.TestDAOCoachGroup;
@@ -9,15 +10,15 @@ import java.util.TreeSet;
 import javax.persistence.*;
 
 @Entity
-@Table(name="education_team")
-public class EducationTeam implements Comparable<EducationTeam>{
-	
-	/**
-     * 
+@Table(name = "education_team")
+public class EducationTeam implements Comparable<EducationTeam>, Serializable {
+
+    /**
+     *
      * Id of the EducationTeam
      */
-	@Id
-	@Column(name="educationTeamID", length=11, nullable=false)
+    @Id
+    @Column(name = "educationTeamID", length = 11, nullable = false)
     private int id;
 
 	@OneToMany(cascade=CascadeType.ALL, targetEntity=CoachGroup.class)
@@ -27,63 +28,80 @@ public class EducationTeam implements Comparable<EducationTeam>{
 	@ManyToMany(cascade=CascadeType.ALL, targetEntity=Teacher.class, fetch=FetchType.EAGER)
 	@JoinColumn(name="bsn")
     private Set<Teacher> teachers = new TreeSet<>();
-
     /**
      * Abbreviation of the EducationTeam; e.g. AO
      */
-    @Column(name="abbreviation", length=64)
+    @Column(name = "abbreviation", length = 64)
     private String abbreviation;
     /**
      * Name of the EducationTeam; e.g. Applicatie Ontwikkelaar
      */
-    @Column(name="name", length=64)
+    @Column(name = "name", length = 64)
     private String name;
-    
-    public EducationTeam() {
-		
-	}
 
-    public void addTeacher(Teacher t){
-		if (teachers.add(t)) {
-			DAOFactory.getTheFactory().getDAOEducationTeam().addTeacherToEducationTeam(t, this);
-		}
-	}
-    
+    public EducationTeam() {
+        this.coachGroups = new TreeSet<>();
+    }
+
+    public EducationTeam(int id, String name) {
+        this.setId(id);
+        this.setName(name);
+        this.coachGroups = new TreeSet<>();
+
+    }
+
+    public EducationTeam(String abbreviation, String name) {
+        this.abbreviation = abbreviation;
+        this.name = name;
+        this.coachGroups = new TreeSet<>();
+
+    }
+
+    public EducationTeam(String abbreviation, String name, int id) {
+        this.setId(id);
+        this.abbreviation = abbreviation;
+        this.name = name;
+        this.coachGroups = new TreeSet<>();
+
+    }
+
+    public void addTeacher(Teacher t) {
+        if (teachers.add(t)) {
+            DAOFactory.getTheFactory().getDAOEducationTeam().addTeacherToEducationTeam(t, this);
+        }
+    }
+
     /**
      * Adds a coachgroup from a EducationTeam and database
+     *
      * @param coachGroup
      */
-    public void addCoachGroup(CoachGroup cg){
-    	cg.setName(this.abbreviation + cg.getName());
-    	this.coachGroups.add(cg);
-    	DAOFactory.getTheFactory().getDAOCoachGroup().create(cg, this);
-    	//TestDAOCoachGroup.getInstance().create(cg);
-    
+    public void addCoachGroup(CoachGroup cg) {
+        cg.setName(this.abbreviation + cg.getName());
+        this.coachGroups.add(cg);
+        DAOFactory.getTheFactory().getDAOCoachGroup().create(cg, this);
     }
+
     /**
      * updates coachgroup from an educationTeam and database
-     * 
+     *
      * @param cg
      * @param oldname
      */
-    public void updateCoachGroup(CoachGroup cg , String oldname){
-    	
-    	//FIXME use DAOFactory.
-    	MYSQLDAOCoachGroup.getInstance().update(cg, this, oldname);
-    	
+    public void updateCoachGroup(CoachGroup cg, String oldname) {
+        DAOFactory.getTheFactory().getDAOCoachGroup().update(cg, oldname);
     }
-    
-    
+
     /**
      * Deletes a coachgroup from a EducationTeam and database
+     *
      * @param CoachGroup coach
      */
-    public void deleteCoachGroup(CoachGroup cg){
-    	this.coachGroups.remove(cg);
-    	DAOFactory.getTheFactory().getDAOCoachGroup().delete(cg);
-    	//TestDAOCoachGroup.getInstance().delete(cg);
+    public void deleteCoachGroup(CoachGroup cg) {
+        this.coachGroups.remove(cg);
+        DAOFactory.getTheFactory().getDAOCoachGroup().delete(cg);
     }
-    
+
     public void updateTeacher(Teacher teacher) {
         for (Teacher ct : teachers) {
             if (teacher.getAbbreviation().equals(ct.getAbbreviation())) {
@@ -98,33 +116,15 @@ public class EducationTeam implements Comparable<EducationTeam>{
             }
         }
     }
-    
-	public void deleteTeacher(Teacher t){
-		//TODO implement a way to fill the teachers collection on startup.
-//		if (teachers.remove(t)) {
-			DAOFactory.getTheFactory().getDAOEducationTeam().deleteTeacherFromEducationTeam(t, this);
-//		}		
-		
-	}
-	
-	public Set<Teacher> getTeachers() {
-		 return teachers;
-	}
-	
-	public EducationTeam(int id, String name) {
-    	this.setId(id);
-    	this.setName(name);
+
+    public void deleteTeacher(Teacher t) {
+        if (teachers.remove(t)) {
+            DAOFactory.getTheFactory().getDAOEducationTeam().deleteTeacherFromEducationTeam(t, this);
+        }
     }
 
-    public EducationTeam(String abbreviation, String name) {
-        this.abbreviation = abbreviation;
-        this.name = name;
-    }
-
-    public EducationTeam(String abbreviation, String name, int id) {
-        this.setId(id);
-        this.abbreviation = abbreviation;
-        this.name = name;
+    public Set<Teacher> getTeachers() {
+        return teachers;
     }
 
     //</editor-fold>
@@ -181,4 +181,30 @@ public class EducationTeam implements Comparable<EducationTeam>{
 	public void loadCoachGroups(){
 		this.coachGroups = DAOFactory.getTheFactory().getDAOCoachGroup().readAllBYTeam(this);
 	}
+
+    @Override
+    public int hashCode() {
+        int hash = 7;
+        hash = 17 * hash + this.id;
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        final EducationTeam other = (EducationTeam) obj;
+        if (this.id != other.id) {
+            return false;
+        }
+        return true;
+    }
+
 }
